# backend/docs/swagger/swagger.yaml
definitions:
...
  internal_users.UserResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_users.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /api/users/{id}:
    get:
      consumes:
      - application/json
      description: Get a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - users
  /api/users/me:
    put:
      consumes:
      - application/json
      description: Update the display name, avatar URL, or status of the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - users
swagger: "2.0"